+++ Calculator/Print.h
@@ -17,7 +17,7 @@ class Print
 public:
     /* 输出原表达式得到的队列 */
     void Myprint (queue<string>s);
-    void printans (int ans);
+    void printans (double ans);
     void print_file (string s1,string s2);
     void print_full_s (string s);
 };
